+++ Formatting/LaMetric.Time.Application.format.ps1
@@ -1,0 +1,1 @@
+Write-FormatView -TypeName LaMetric.Time.Application -Property title, vendor, version