+++ Types/LaMetric.Icon/get_BigThumbnail.ps1
@@ -1,0 +1,3 @@
+[uri]("https://developer.lametric.com/content/apps/icon_thumbs/$(
+    $this.IconID -replace '\D'
+)_icon_thumb_big.png")